# Daily Perf Improver: Simplex Theory Solver Optimizations (Round 2)