@dsyme 2025-09-17
The perf evidence ran ok but the numbers didn't hold up once we multiplied number of iterations x100.  

```
Column size: 100000 elements, Iterations: 5000
Original: 723106 μs
Optimized: 765466 μs
Speedup: 0.944661x
```